refactor(core): decouple core layer from Preact framework (IEX-2294)
Move Preact-specific providers/hooks into preact/ subdirectories;
originals become re-export shims for zero consumer import churn.

- CoreContext, AnalyticsContext, ConfigContext adapters
- Replace Preact RefObject with generic ComponentRef interface
- Make interpolateElement framework-agnostic (generic type param)
- Add Vue composable stubs as proof-of-concept
