+++ src/core/Context/preact/useCoreContext.ts
@@ -1,0 +1,6 @@
+import { useContext } from 'preact/hooks';
+import { CoreContext } from '../CoreContext';
+
+const useCoreContext = () => useContext(CoreContext);
+
+export default useCoreContext;
