+++ src/core/Context/types.ts
@@ -1,7 +1,10 @@
 import Localization from '../Localization';
 import { onErrorHandler } from '../types';
-import { RefObject } from 'preact';
 import { AssetOptions } from '../Assets/Assets';
+
+export interface ComponentRef<T = HTMLDivElement> {
+    current: T | null;
+}
 
 export interface CommonPropsTypes {
     isCollatingErrors?: boolean;
@@ -14,7 +17,7 @@ export interface CoreProviderProps {
     loadingContext?: string;
     updateCore?: () => void;
     externalErrorHandler?: onErrorHandler | null;
-    componentRef: RefObject<HTMLDivElement>;
+    componentRef: ComponentRef<HTMLDivElement>;
     getImageAsset?: (props: AssetOptions) => string;
     getDatasetAsset?: (props: AssetOptions) => string;
     getCdnConfig?: <Fallback>(props: { name: string; extension?: string; subFolder?: string; fallback?: Fallback }) => Promise<Fallback>;
