+++ src/core/Context/useCoreContext.ts
@@ -1,6 +1,1 @@
-import { useContext } from 'preact/hooks';
-import { CoreContext } from './CoreContext';
-
-const useCoreContext = () => useContext(CoreContext);
-
-export default useCoreContext;
+export { default } from './preact/useCoreContext';
